Fix CSS for better scrolling
The git gist had the problem that it would only render anything in the
viewport and not overflowed. This forced me to switch to a better style
scheme in which the nav bar is affixed by position fixed and the glass
div is free to scroll infinitely. This has also allowed me to remove the
ground div entirely.