Adjust the max radius calculation for area
The area is a better heuristic for the largest that the radius could be.
The aspect ratio is only applicable if one cares about the height, but
if the width will be smaller than the height, this is a bad measure.

The square root of the area is a better amount to multiply by because it
gives the measure of how much space one actually has to work with.